finished DiscordClass config update
This finished the code rewrite needed, and code cleanup, for the discord bot to use the new YAML file. This should allow it to have a new text group added in config file, then look for messages mentioning the bot in the corresponding  discord channel. This should work the same as before, just you will be able to add new groups through the config file. Have not tested this yet, but the initial code is finished.