compose file updated for this branch and added todo
I updated the docker compose file to work for this branch in hopes to test the docker files tonight. Unfortunately I forgot to program the add_subscriber in NocoClass to deal with if a number is already in the database. Currently it would work, but would either reset the update group column to the most recent group they signed up for, or it will create a new line. Pretty sure it created the new line but im too tired to figure it out. Need to deal with if subscriber is already signed up for one of the groups, and if so they just need the tag added instead of creating another entry in the database.